Rename `prefixer` mixin to `animate-prefixer`
Bourbon, the simple and lightweight mixin library for Sass also
implements a prefixer() mixin. When importing this library after
Bourbon, many things break. A quick workaround is to just import this
library before Bourbon, but in the long run it would be best to avoid
this naming collision.

This commit resolves this by prefixing the name of the `prefixer` mixin
with `animate-`. This name should be unique enough to reduce the
likelihood of this being a problem.

Some trailing space removal came along for the ride.

Fixes #5